Include MD5 hash of absolute path in temporary folder name
Running parallel instances of the script on a directory tree containing
some duplicate base filenames will cause files to be dropped in the same
temporary folder, causing all sorts of errors.  Best to guarantee unique
names for temporary folders.

MD5 is of course insecure, but it’s a nice, short hash with enough
strength to block an unintended collision.